+++ fiboa_cli/convert_utils.py
@@ -1,10 +1,10 @@
 from .const import STAC_TABLE_EXTENSION
 from .version import fiboa_version
-from .util import log, get_fs, to_iso8601
+from .util import log, get_fs, name_from_uri, to_iso8601
 from .parquet import create_parquet
 
-from urllib.parse import urlparse
 from tempfile import TemporaryDirectory
+from shapely.geometry import box
 
 import os
 import re
@@ -18,7 +18,9 @@ import py7zr
 def convert(
         output_file, cache_path,
         urls, columns,
-        id, title, description, bbox,
+        id, title, description,
+        input_files = None,
+        bbox = None,
         provider_name = None,
         provider_url = None,
         source_coop_url = None,
@@ -28,6 +30,7 @@ def convert(
         column_filters = {},
         column_migrations = {},
         migration = None,
+        file_migration = None,
         attribution = None,
         store_collection = False,
         license = None,
@@ -37,20 +40,33 @@ def convert(
     """
     Converts a field boundary datasets to fiboa.
     """
-    if len(bbox) != 4:
-        raise ValueError("Bounding box must be of length 4")
+    if bbox is not None and len(bbox) != 4:
+        raise ValueError("If provided, the bounding box must consist of 4 numbers")
 
-    log(f"Getting file(s) if not cached yet")
+    if input_files is not None and isinstance(input_files, dict) and len(input_files) > 0:
+        log("Using user provided input file(s) instead of the pre-defined file(s)", "warning")
+        urls = input_files
+    elif urls is None:
+        raise ValueError("No input files provided")
+
+    log("Getting file(s) if not cached yet")
     paths = download_files(urls, cache_path)
 
-    log(f"Reading into GeoDataFrame")
+    log("Reading into GeoDataFrame")
     gdfs = []
-    for path in paths:
+    for path, uri in paths:
         # If file is a parquet file then read with read_parquet
         if path.endswith(".parquet") or path.endswith(".geoparquet"):
             data = gpd.read_parquet(path, **kwargs)
         else:
             data = gpd.read_file(path, **kwargs)
+
+        # 0. Run migration per file
+        if callable(file_migration):
+            log("Applying per file migrations")
+            data = file_migration(data, path, uri)
+            if not isinstance(data, gpd.GeoDataFrame):
+                raise ValueError("Per-file migration function must return a GeoDataFrame")
 
         gdfs.append(data)
 
@@ -173,7 +189,7 @@ def convert(
 
 def create_collection(
         gdf,
-        id, title, description, bbox,
+        id, title, description, bbox = None,
         provider_name = None,
         provider_url = None,
         source_coop_url = None,
@@ -184,6 +200,9 @@ def create_collection(
     """
     Creates a collection for the field boundary datasets.
     """
+    if bbox is None:
+        bbox = list(gpd.GeoSeries([box(*gdf.total_bounds)], crs=gdf.crs).to_crs(epsg=4326).total_bounds)
+
     collection = {
         "fiboa_version": fiboa_version,
         "fiboa_extensions": extensions,
@@ -196,7 +215,7 @@ def create_collection(
         "extent": {
             "spatial": {
                 "bbox": [bbox]
-            }
+            },
         },
         "links": []
     }
@@ -304,12 +323,12 @@ def download_files(uris, cache_folder = None):
     if cache_folder is None:
         args = {}
         if sys.version_info.major >= 3 and sys.version_info.minor >= 12:
-            args.delete = False # only available in Python 3.12 and later
+            args["delete"] = False # only available in Python 3.12 and later
         with TemporaryDirectory(**args) as tmp_folder:
             cache_folder = tmp_folder
 
     if isinstance(uris, str):
-        uris = {uris: name_from_url(uris)}
+        uris = {uris: name_from_uri(uris)}
 
     paths = []
     i = 0
@@ -318,7 +337,7 @@ def download_files(uris, cache_folder = None):
         is_archive = isinstance(target, list)
         if is_archive:
             try:
-                name = name_from_url(uri)
+                name = name_from_uri(uri)
                 # if there's no file extension, it's likely a folder, which may not be unique
                 if "." not in name:
                     name = str(i)
@@ -352,15 +371,11 @@ def download_files(uris, cache_folder = None):
 
         if is_archive:
             for filename in target:
-                paths.append(os.path.join(zip_folder, filename))
+                paths.append((os.path.join(zip_folder, filename), uri))
         else:
-            paths.append(cache_file)
+            paths.append((cache_file, uri))
 
     return paths
-
-
-def name_from_url(url):
-    return os.path.basename(urlparse(url).path)
 
 
 def stream_file(fs, src_uri, dst_file, chunk_size = 10 * 1024 * 1024):