+++ fiboa_cli/datasets/de_th.py
@@ -6,6 +6,7 @@ import pandas as pd
 SOURCES = "https://www.geoproxy.geoportal-th.de/download-service/opendata/agrar/DGK_Thue.zip"
 
 ID = "de_th"
+SHORT_NAME = "Germany, Thuringia"
 TITLE = "Field boundaries for Thuringia, Germany"
 DESCRIPTION = """For use in the application procedure of the Integrated Administration and Control System (IACS), digital data layers are required that represent the current situation of agricultural use with the required accuracy. The field block is a contiguous agricultural area of one or more farmers surrounded by permanent boundaries. The field block thus contains information on the geographical location of the outer boundaries of the agricultural area. Reference parcels are uniquely numbered throughout Germany (Feldblockident - FBI). They also have a field block size (maximum eligible area) and a land use category.
 
@@ -21,9 +22,6 @@ The field blocks are classified separately according to the main land uses of ar
 Landscape elements (LE) are considered part of the eligible agricultural area under defined conditions. In Thuringia, these permanent conditional features are designated as a separate field block (FB) and are therefore part of the Thuringian area reference system (field block reference). They must have a clear reference to an UAA (agricultural land), i.e. they are located within an arable, permanent grassland or permanent crop area or border directly on it.
 
 To produce the DGK-Lw, (official) orthophotos from the Thuringian Land Registry and Surveying Administration (TLBG) and orthophotos from the TLLLR's own aerial surveys are interpreted. The origin of this image data is 50% of the state area each year, so that up-to-date image data is available for the entire Thuringian state area every year."""
-
-# Taken from http://osmtipps.lefty1963.de/2008/10/bundeslnder.html
-BBOX = [9.8778443239,50.2042330625,12.6531964048,51.6490678544]
 
 PROVIDER_NAME = "Thüringer Landesamt für Landwirtschaft und Ländlichen Raum"
 PROVIDER_URL = "https://geomis.geoportal-th.de/geonetwork/srv/ger/catalog.search#/metadata/D872F2D6-60BC-11D6-B67D-00E0290F5BA0"
@@ -121,7 +119,7 @@ MISSING_SCHEMAS = {
 
 
 # Conversion function, usually no changes required
-def convert(output_file, cache = None, source_coop_url = None, collection = False, compression = None):
+def convert(output_file, input_files = None, cache = None, source_coop_url = None, collection = False, compression = None):
     convert_(
         output_file,
         cache,
@@ -130,7 +128,7 @@ def convert(output_file, cache = None, source_coop_url = None, collection = Fals
         ID,
         TITLE,
         DESCRIPTION,
-        BBOX,
+        input_files=input_files,
         migration=MIGRATION,
         provider_name=PROVIDER_NAME,
         provider_url=PROVIDER_URL,