+++ fiboa_cli/datasets/fi.py
@@ -3,13 +3,13 @@ import pandas as pd
 
 SOURCES = "https://download.inspire.ruokavirasto-awsa.com/data/2023/LandUse.ExistingLandUse.GSAAAgriculturalParcel.gpkg"
 ID = "fi"
+SHORT_NAME = "Finland"
 TITLE = "Finnish Crop Fields (Maatalousmaa)"
 DESCRIPTION = """
 The Finnish Food Authority (FFA) since 2020 produces spatial data sets,
 more specifically in this context the "Field parcel register" and "Agricultural parcel containing spatial data".
 A set called "Agricultural land: arable land, permanent grassland or permanent crop (land use)".
 """
-BBOX = [16.23255510557983, 59.6677095773865, 32.81781081278198, 69.99103510928764]
 PROVIDER_NAME = "Finnish Food Authority"
 PROVIDER_URL = "https://www.ruokavirasto.fi/en/about-us/open-information/spatial-data-sets/"
 ATTRIBUTION = "Finnish Food Authority"
@@ -43,7 +43,7 @@ MISSING_SCHEMAS = {
     }
 }
 
-def convert(output_file, cache = None, source_coop_url = None, collection = False, compression = None):
+def convert(output_file, input_files = None, cache = None, source_coop_url = None, collection = False, compression = None):
     convert_(
         output_file,
         cache,
@@ -52,7 +52,7 @@ def convert(output_file, cache = None, source_coop_url = None, collection = Fals
         ID,
         TITLE,
         DESCRIPTION,
-        BBOX,
+        input_files=input_files,
         provider_name=PROVIDER_NAME,
         provider_url=PROVIDER_URL,
         source_coop_url=source_coop_url,